> bug: P with valency four
There ar enumber of cases with P with valency four. Some of them seem to be correct, but suspicious (e.g. 3JE, 4NE)
In one particular case - DG9 there must be four oxygens around P2A and after the link formation between SER and this residue 
one of the oxygens is removed. I added corrected DG9 and corresponding link to the monomer library.

> reported by: Marcin Wojdyr